fixed version key problem

+++ src/main/scala/org/ergoplatform/nodeView/history/ErgoHistory.scala
@@ -9,7 +9,7 @@ import org.ergoplatform.modifiers.history._
 import org.ergoplatform.modifiers.history.header.{Header, PreGenesisHeader}
 import org.ergoplatform.modifiers.{BlockSection, ErgoFullBlock, NonHeaderBlockSection}
 import org.ergoplatform.nodeView.history.extra.ExtraIndexer.ReceivableMessages.StartExtraIndexer
-import org.ergoplatform.nodeView.history.extra.ExtraIndexer.{NewestVersion, NewestVersionBytes, SchemaVersionKey, getIndex}
+import org.ergoplatform.nodeView.history.extra.ExtraIndexer.{IndexedHeightKey, NewestVersion, NewestVersionBytes, SchemaVersionKey, getIndex}
 import org.ergoplatform.nodeView.history.storage.HistoryStorage
 import org.ergoplatform.nodeView.history.storage.modifierprocessors._
 import org.ergoplatform.settings._
@@ -286,8 +286,9 @@ object ErgoHistory extends ScorexLogging {
 
     if(ergoSettings.nodeSettings.extraIndex) { // check extra indexer db schema
       val schemaVersion: Int = getIndex(SchemaVersionKey, db).getInt
-      if (schemaVersion != NewestVersion) { // older schema -> delete and reopen db
-        db = db.deleteExtraDB(ergoSettings)
+      if (schemaVersion != NewestVersion) {
+        if(getIndex(IndexedHeightKey, db).getInt > 0)
+          db = db.deleteExtraDB(ergoSettings) // older schema -> delete and reopen db
         db.insertExtra(Array((SchemaVersionKey, NewestVersionBytes)), Array.empty) // update version key
       }
     }
